+++ tests.py
@@ -459,12 +459,19 @@ class StreamTestCase(unittest.TestCase):
         assert list(s) == [1,2,3,4,5]
 
     def test_lazy_slicing(self):
-        s = Stream() << xrange(10)
-        self.assertEqual(len(s._collection), 0)
+        s = Stream() << iters.range(10)
+        self.assertEqual(s.cursor(), 0)
+
         s_slice = s[:5]
-        self.assertEqual(len(s._collection), 0)
+        self.assertEqual(s.cursor(), 0)
         self.assertEqual(len(list(s_slice)), 5)
-        # not sure if it's worth testing negative slices, as they cannot be lazy
+
+    def test_lazy_slicing_recursive(self):
+        s = Stream() << iters.range(10)
+        sf = s[1:3][0:2]
+
+        self.assertEqual(s.cursor(), 0)
+        self.assertEqual(len(list(sf)), 2)
 
     def test_fib_infinite_stream(self):
         from operator import add 
@@ -475,6 +482,9 @@ class StreamTestCase(unittest.TestCase):
         self.assertEqual([0,1,1,2,3,5,8,13,21,34], list(iters.take(10, fib)))
         self.assertEqual(6765, fib[20])
         self.assertEqual([832040,1346269,2178309,3524578,5702887], list(fib[30:35]))
+        # 35 elements should be already evaluated
+        self.assertEqual(fib.cursor(), 35)
+
 
 if __name__ == '__main__':
     unittest.main()